+++ src/dtos/PillInfoDto.js
