+++ src/dtos/response/PillSearchDto.js
