+++ src/models/Pill.js
